refactor(app): move services into lib folder
