+++ packages/app/src/FileDropzone.tsx
@@ -3,7 +3,7 @@ import { useDropzone } from 'react-dropzone'
 import { useTranslation } from 'react-i18next'
 import { LuUpload } from 'react-icons/lu'
 
-import { openFromFile } from './services/fileService'
+import { openFromFile } from './lib/file-handler'
 import { useDialogStore } from './stores/dialogStore'
 import { cn } from './utils'
