+++ packages/app/src/components/ShortcutHandler.tsx
@@ -1,9 +1,9 @@
 import { type FC, useEffect, useMemo } from 'react'
 
-import { toggleGroup } from '@/services/actions'
-import { openFileFromUserSelect, saveToFile } from '@/services/fileService'
-import { getLogger } from '@/services/loggerService'
-import type { ShortcutActionsEnum } from '@/services/settings'
+import { toggleGroup } from '@/lib/actions'
+import { openFileFromUserSelect, saveToFile } from '@/lib/file-handler'
+import { getLogger } from '@/lib/logger'
+import type { ShortcutActionsEnum } from '@/lib/settings'
 import { useDialogStore } from '@/stores/dialogStore'
 import { useDisplayEntityStore } from '@/stores/displayEntityStore'
 import { useEditorStore } from '@/stores/editorStore'
