+++ packages/app/src/components/dialog/settings/AppearancePage.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useShallow } from 'zustand/shallow'
 
-import type { Settings } from '@/services/settings'
+import type { Settings } from '@/lib/settings'
 import { useEditorStore } from '@/stores/editorStore'
 
 const AppearancePage: FC = () => {
