+++ packages/app/src/components/sidebar/TransformsPanel.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { MathUtils } from 'three'
 import { useShallow } from 'zustand/shallow'
 
-import { getLogger } from '@/services/loggerService'
+import { getLogger } from '@/lib/logger'
 import { useDisplayEntityStore } from '@/stores/displayEntityStore'
 import { useEditorStore } from '@/stores/editorStore'
 import { useEntityRefStore } from '@/stores/entityRefStore'
