+++ packages/app/src/lib/resources/textMesh.ts
