+++ packages/app/src/lib/settings.ts
