+++ packages/app/src/stores/cacheStore.ts
@@ -4,8 +4,8 @@ import { create } from 'zustand'
 import { immer } from 'zustand/middleware/immer'
 
 import { CDNBaseUrl } from '@/constants'
-import { getLogger } from '@/services/loggerService'
-import { loadModel } from '@/services/resources/model'
+import { getLogger } from '@/lib/logger'
+import { loadModel } from '@/lib/resources/model'
 import type {
   AssetFileInfos,
   BlockstatesData,
